+++ src/Views/Columns/BooleanColumn.php
@@ -13,6 +13,7 @@ class BooleanColumn extends Column
     use BooleanColumnConfiguration,
         BooleanColumnHelpers;
 
+    protected string $type = 'icons';
     protected bool $successValue = true;
     protected string $view = 'livewire-tables::includes.columns.boolean';
     protected $callback;
@@ -28,6 +29,7 @@ class BooleanColumn extends Column
         return view($this->getView())
             ->withComponent($this->getComponent())
             ->withSuccessValue($this->getSuccessValue())
+            ->withType($this->getType())
             ->withStatus($this->hasCallback() ? call_user_func($this->getCallback(), $value, $row) : (bool)$value === true);
     }
 }