+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to `laravel-livewire-tables` will be documented in this file
 
 ## [Unreleased]
 
+### Added
+
+- Added functionality to display BooleanColumn as Yes/No instead of icons.
+
 ## [2.5.0] - 2022-05-03
 
 ### Added